update/added_protocol_version_in_status
Update EraID to uint64

+++ rpc/response.go
@@ -511,7 +511,7 @@ func (b QueryBalanceDetailsResult) GetRawJSON() json.RawMessage {
 type InfoGetRewardResult struct {
 	APIVersion     string          `json:"api_version"`
 	DelegationRate float32         `json:"delegation_rate"`
-	EraID          uint32          `json:"era_id"`
+	EraID          uint64          `json:"era_id"`
 	RewardAmount   clvalue.UInt512 `json:"reward_amount"`
 
 	rawJSON json.RawMessage
