+++ types/minimal_block_info.go
@@ -9,7 +9,7 @@ import (
 
 type MinimalBlockInfo struct {
 	Creator       keypair.PublicKey `json:"creator"`
-	EraID         uint32            `json:"era_id"`
+	EraID         uint64            `json:"era_id"`
 	Hash          key.Hash          `json:"hash"`
 	Height        uint32            `json:"height"`
 	StateRootHash key.Hash          `json:"state_root_hash"`
